Lower eql-specializer overhead plus put its name slot to good use
If method-table-eql-specializers is t, set it to a cache with eql-specializers by parameter. This reduces the method-table-method-list's size eliminating unnecessarily restricted cached effective method functions i.e. for any given argument, the old code searches for a corresponding singleton which may not be relevant to that parameter or indeed it may not even pertain to the given generic function. We also use class-of, and gethash when necessary, only once per argument. Using conses for eql-specializers hopefully gives faster cache access.

Generate eql-specializers for (object intern-form) pairs such that if two such classes have eql objects then one must be a subclass of the other. In this way they satisfy method applicability and also can be named to preserve their intern-form for debugging and inspecting purposes.